alfa version -refactor
