+++ src/main/java/pages/HhPages/ProfileHhPage.java
@@ -1,4 +1,4 @@
-package pages.HhClasses;
+package pages.HhPages;
 
 import org.openqa.selenium.StaleElementReferenceException;
 import org.openqa.selenium.TimeoutException;
