+++ src/main/java/pages/HhPages/SentCvPopUp.java
@@ -1,6 +1,5 @@
-package pages.HhClasses;
+package pages.HhPages;
 
-import org.openqa.selenium.JavascriptExecutor;
 import org.openqa.selenium.StaleElementReferenceException;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
